Split `handleNotificationsMiddleware` in `notificationsMiddleware` and `hashtagsMiddleware`
