+++ backend/src/middleware/index.js
@@ -12,7 +12,8 @@ import user from './userMiddleware'
 import includedFields from './includedFieldsMiddleware'
 import orderBy from './orderByMiddleware'
 import validation from './validation/validationMiddleware'
-import handleNotifications from './handleNotifications/handleNotificationsMiddleware'
+import notifications from './notifications/notificationsMiddleware'
+import hashtags from './hashtags/hashtagsMiddleware'
 import email from './email/emailMiddleware'
 import sentry from './sentryMiddleware'
 
@@ -25,13 +26,16 @@ export default schema => {
     validation,
     sluggify,
     excerpt,
-    handleNotifications,
+    notifications,
+    hashtags,
     xss,
     softDelete,
     user,
     includedFields,
     orderBy,
-    email: email({ isEnabled: CONFIG.SMTP_HOST && CONFIG.SMTP_PORT }),
+    email: email({
+      isEnabled: CONFIG.SMTP_HOST && CONFIG.SMTP_PORT,
+    }),
   }
 
   let order = [
@@ -43,7 +47,8 @@ export default schema => {
     'sluggify',
     'excerpt',
     'email',
-    'handleNotifications',
+    'notifications',
+    'hashtags',
     'xss',
     'softDelete',
     'user',
