+++ backend/src/middleware/notifications/notificationsMiddleware.spec.js
@@ -461,112 +461,3 @@ describe('notifications', () => {
     })
   })
 })
-
-describe('Hashtags', () => {
-  const id = 'p135'
-  const title = 'Two Hashtags'
-  const postContent =
-    '<p>Hey Dude, <a class="hashtag" href="/search/hashtag/Democracy">#Democracy</a> should work equal for everybody!? That seems to be the only way to have equal <a class="hashtag" href="/search/hashtag/Liberty">#Liberty</a> for everyone.</p>'
-  const postWithHastagsQuery = gql`
-    query($id: ID) {
-      Post(id: $id) {
-        tags {
-          id
-        }
-      }
-    }
-  `
-  const postWithHastagsVariables = {
-    id,
-  }
-
-  describe('authenticated', () => {
-    beforeEach(async () => {
-      authenticatedUser = await notifiedUser.toJson()
-    })
-
-    describe('create a Post with Hashtags', () => {
-      beforeEach(async () => {
-        await mutate({
-          mutation: createPostMutation,
-          variables: {
-            id,
-            title,
-            postContent,
-            categoryIds,
-          },
-        })
-      })
-
-      it('both Hashtags are created with the "id" set to their "name"', async () => {
-        const expected = [
-          {
-            id: 'Democracy',
-          },
-          {
-            id: 'Liberty',
-          },
-        ]
-        await expect(
-          query({
-            query: postWithHastagsQuery,
-            variables: postWithHastagsVariables,
-          }),
-        ).resolves.toEqual(
-          expect.objectContaining({
-            data: {
-              Post: [
-                {
-                  tags: expect.arrayContaining(expected),
-                },
-              ],
-            },
-          }),
-        )
-      })
-
-      describe('afterwards update the Post by removing a Hashtag, leaving a Hashtag and add a Hashtag', () => {
-        // The already existing Hashtag has no class at this point.
-        const postContent =
-          '<p>Hey Dude, <a class="hashtag" href="/search/hashtag/Elections">#Elections</a> should work equal for everybody!? That seems to be the only way to have equal <a href="/search/hashtag/Liberty">#Liberty</a> for everyone.</p>'
-
-        it('only one previous Hashtag and the new Hashtag exists', async () => {
-          await mutate({
-            mutation: updatePostMutation,
-            variables: {
-              id,
-              title,
-              postContent,
-              categoryIds,
-            },
-          })
-
-          const expected = [
-            {
-              id: 'Elections',
-            },
-            {
-              id: 'Liberty',
-            },
-          ]
-          await expect(
-            query({
-              query: postWithHastagsQuery,
-              variables: postWithHastagsVariables,
-            }),
-          ).resolves.toEqual(
-            expect.objectContaining({
-              data: {
-                Post: [
-                  {
-                    tags: expect.arrayContaining(expected),
-                  },
-                ],
-              },
-            }),
-          )
-        })
-      })
-    })
-  })
-})
